Reformatted tests
