+++ tests/conftest.py
@@ -21,9 +21,12 @@ If not, see <https://www.gnu.org/licenses/>.
 """
 
 import os
-import weakref
+import time
+import multiprocessing
+import rpyc
 import pytest
 from PySide2 import QtWidgets
+from PySide2.QtCore import QTimer
 from qudi.core import application
 from qudi.util.yaml import yaml_load
 
@@ -43,8 +46,7 @@ def qudi_instance():
     instance = application.Qudi.instance()
     if instance is None:
         instance = application.Qudi(config_file=CONFIG)
-    instance_weak = weakref.ref(instance)
-    return instance_weak()
+    return instance
 
 @pytest.fixture(scope="module")
 def module_manager(qudi_instance):
@@ -72,3 +74,38 @@ def hardware_modules(config):
     base = 'hardware'
     modules  = list(config[base].keys())
     return modules
+
+def run_qudi(timeout=150000):
+    """
+    This function runs a qudi instance with a timer
+    """    
+    app_cls = QtWidgets.QApplication
+    app = app_cls.instance()
+    if app is None:
+        app = app_cls()
+    qudi_instance = application.Qudi.instance()
+    if qudi_instance is None:
+        qudi_instance = application.Qudi(config_file=CONFIG)
+    QTimer.singleShot(timeout, qudi_instance.quit)
+    qudi_instance.run()
+
+
+@pytest.fixture(scope='module')
+def start_qudi_process():
+    """This fixture starts the Qudi process and ensures it's running before tests."""
+    qudi_process = multiprocessing.Process(target=run_qudi)
+    qudi_process.start()
+    time.sleep(5)
+    yield
+    qudi_process.join(timeout=10)
+    if qudi_process.is_alive():
+        qudi_process.terminate()
+
+@pytest.fixture(scope='module')
+def remote_instance(start_qudi_process):
+    """This fixture connects the running qudi server through rpyc client and returns client instance"""
+    time.sleep(5)
+    conn = rpyc.connect("localhost", 18861, config={'sync_request_timeout': 60})
+    root = conn.root
+    qudi_instance = root._qudi
+    return qudi_instance
