+++ tests/test_OdmrLogic.py
@@ -26,34 +26,14 @@ import math
 import numpy as np
 import coverage
 import pytest
-import rpyc
-import multiprocessing
-from PySide2 import QtWidgets
-from PySide2.QtCore import QTimer
-from qudi.core import application
 from qudi.util.network import netobtain
 
-CONFIG = os.path.join(os.getcwd(),'tests/test.cfg')
 MODULE = 'odmr_logic'
 BASE = 'logic'
 CHANNELS = ('APD counts', 'Photodiode')
 FIT_MODEL = 'Gaussian Dip'
 TOLERANCE = 10 # tolerance for signal data range
 
-
-def run_qudi():
-    """
-    This function runs a qudi instance with a timer
-    """    
-    app_cls = QtWidgets.QApplication
-    app = app_cls.instance()
-    if app is None:
-        app = app_cls()
-    qudi_instance = application.Qudi.instance()
-    if qudi_instance is None:
-        qudi_instance = application.Qudi(config_file=CONFIG)
-    QTimer.singleShot(50000, qudi_instance.quit)
-    qudi_instance.run()
 
 def get_scanner(module):
     return module._data_scanner()
@@ -85,29 +65,16 @@ def get_tolerance(value, bound):
     return int(value + value * TOLERANCE/100) if bound == 'upper' else int(value - value * TOLERANCE/100)
 
 
-@pytest.fixture(scope='module', autouse=True)
-def start_qudi_process():
-    """This fixture starts the Qudi process and ensures it's running before tests."""
-    qudi_process = multiprocessing.Process(target=run_qudi)
-    qudi_process.start()
-    time.sleep(10)
-    yield
-    qudi_process.join(timeout=10)
-    if qudi_process.is_alive():
-        qudi_process.terminate()
-
-@pytest.fixture(scope='module')
-def remote_instance():
-    """This fixture connects the running qudi server through rpyc client and returns client instance"""
-    time.sleep(10)
-    conn = rpyc.connect("localhost", 18861, config={'sync_request_timeout': 60})
-    root = conn.root
-    qudi_instance = root._qudi
-    return qudi_instance
-
 @pytest.fixture(scope='module')
 def module(remote_instance):
-    """This fixture return Odmr logic instance"""
+    """ ""
+    This fixture return Odmr logic instance
+
+    Parameters
+    ----------
+    remote_instance : fixture
+        Remote qudi instance
+    """
     module_manager = remote_instance.module_manager
     odmr_gui = 'odmr_gui'
     odmr_logic = 'odmr_logic'
