+++ tests/test_modify_status_var_file.py
@@ -20,23 +20,17 @@ You should have received a copy of the GNU Lesser General Public License along w
 If not, see <https://www.gnu.org/licenses/>.
 """
 
-import os
 import random
 import time
-import multiprocessing
-import pytest
 import string
+import pytest
 import numpy as np
-import rpyc
-from PySide2 import QtWidgets
-from PySide2.QtCore import QTimer
-from qudi.core import application
 from qudi.util.yaml import yaml_load, yaml_dump
 from qudi.util.paths import get_module_app_data_path
 
-
-CONFIG = os.path.join(os.getcwd(),'tests/test.cfg')
-
+@pytest.fixture
+def module_manager(remote_instance):
+    return remote_instance.module_manager
 
 def generate_random_value(var):
     if isinstance(var, int):
@@ -93,42 +87,6 @@ def dump_status_variables(vars, file_path):
     except Exception as e:
         print("Failed to save status variables:", e)
 
-def run_qudi():
-    """
-    This function runs a qudi instance with a timer
-    """    
-    app_cls = QtWidgets.QApplication
-    app = app_cls.instance()
-    if app is None:
-        app = app_cls()
-    qudi_instance = application.Qudi.instance()
-    if qudi_instance is None:
-        qudi_instance = application.Qudi(config_file=CONFIG)
-    QTimer.singleShot(150000, qudi_instance.quit)
-    qudi_instance.run()
-
-
-@pytest.fixture(scope='module')
-def start_qudi_process():
-    """This fixture starts the Qudi process and ensures it's running before tests."""
-    qudi_process = multiprocessing.Process(target=run_qudi)
-    qudi_process.start()
-    time.sleep(10)
-    yield
-    qudi_process.join(timeout=10)
-
-    if qudi_process.is_alive():
-        qudi_process.terminate()
-
-@pytest.fixture(scope='module')
-def remote_instance(start_qudi_process):
-    """This fixture connects the running qudi server through rpyc client and returns client instance"""
-    time.sleep(10)
-    conn = rpyc.connect("localhost", 18861, config={'sync_request_timeout': 60})
-    root = conn.root
-    qudi_instance = root._qudi
-    return qudi_instance
-
 
 '''
 # Updating status vars to illegal values might break modules, ( For example, PID has no way no reset status vars)
@@ -178,15 +136,20 @@ def test_update_status_vars(qudi_instance, gui_modules, hardware_modules, qt_app
             dump_status_variables(modified_vars, status_var_file_path)
 '''
 
-def test_status_vars(remote_instance, gui_modules, qudi_instance):
+def test_status_vars(module_manager, gui_modules, qudi_instance, qt_app):
     """Test if GUI modules are activated correctly after updating the saved files.
 
     Parameters
     ----------
-    remote_instance : fixture
-        Remote qudi instance
+    module_manager : fixture
+        Remote module manager
+    gui_modules : fixture
+        List of GUI modules
+    qudi_instance : fixture
+        So that Qudi objects don't go out of scope
+    qt_app : fixture
+        Instance of Qt app
     """    
-    module_manager = remote_instance.module_manager
     for gui_module in gui_modules:
         module_manager.activate_module(gui_module)
         gui_managed_module = module_manager.modules[gui_module]
@@ -194,5 +157,5 @@ def test_status_vars(remote_instance, gui_modules, qudi_instance):
         time.sleep(5)
         module_manager.deactivate_module(gui_module)
         time.sleep(1)
-            
+    qudi_instance.quit()
